Se agrego la funcion de eliminar

+++ claseCatalogos.php
@@ -39,7 +39,7 @@ class catalogos{
    		$archivos=explode(",,,",$archivos);
    		$mensaje="";
    		for($i=0;$i<count($archivos);$i++){
-			echo $sql="SELECT CAT2_ARCHIVO.ID_ARCHIVO AS ID_ARCHIVO,PATH FROM CAT2_ARCHIVO INNER JOIN CAT2_ARCHIVO_USUARIO ON CAT2_ARCHIVO.ID_ARCHIVO=CAT2_ARCHIVO_USUARIO.ID_ARCHIVO WHERE CAT2_ARCHIVO.ID_ARCHIVO='".$archivos[$i]."'";
+			$sql="SELECT CAT2_ARCHIVO.ID_ARCHIVO AS ID_ARCHIVO,PATH,ID_USUARIO_ASIGNADO FROM CAT2_ARCHIVO INNER JOIN CAT2_ARCHIVO_USUARIO ON CAT2_ARCHIVO.ID_ARCHIVO=CAT2_ARCHIVO_USUARIO.ID_ARCHIVO WHERE CAT2_ARCHIVO.ID_ARCHIVO='".$archivos[$i]."'";
    			$res=$objDb->sqlQuery($sql);
 
    			while($row=$objDb->sqlFetchArray($res)){
@@ -50,23 +50,28 @@ class catalogos{
 	   				$usuariosPrevios=$usuariosPrevios.",".$row["ID_USUARIO_ASIGNADO"];
 	   			}
    			}
-			//se procede a eliminar el arciv
-   			$sql="DELETE FROM CAT2_ARCHIVO_USUARIO WHERE ID_ARCHIVO='".$archivos[$i]."'";
 
-   			echo "<br>".$path;
+			if(unlink($path)){//se procede a eliminar el archivo
+				//se procede a eliminar el arciv
+	   			$sql="DELETE FROM CAT2_ARCHIVO_USUARIO WHERE ID_ARCHIVO='".$archivos[$i]."'";
+	   			
+	   			$res=$objDb->sqlQuery($sql);
+
+	   			if($res){
+	   				$sql1="DELETE FROM CAT2_ARCHIVO WHERE ID_ARCHIVO='".$archivos[$i]."'";
+	   				$res1=$objDb->sqlQuery($sql1);
+	   				if($res1){
+	   					$mensaje=1;//echo "<br>A borrar";		
+	   				}else{
+	   					$mensaje=0;
+	   				}
+	   			}
+		   	}else{
+		   		$mensaje=0;//echo "<br>Archivo no encontrado";
+		   	}
 
 
-   			/*$res=$objDb->sqlQuery($sql);
-   			if($res){
-   				$sql1="DELEte FROM CAT2_ARCHIVO WHERE ID_ARCHIVO='".$archivos[$i]."'";
-   				$res1=$objDb->sqlQuery($sql1);
-   				if($res1){
-   					$mensaje=1;
-   					//se procede a eliminar el archivo
-   				}else{
-   					$mensaje=0;
-   				}
-   			}*/
+
    		}
    		$mensaje=$mensaje."|||".$usuariosPrevios;
    		return $mensaje;
@@ -97,12 +102,8 @@ class catalogos{
 	*/
 	public function creaNuevoCatalogo($idCliente,$nombre){
 		$objDb=$this->iniciarConexionDb();
-		$sqlE="INSERT INTO CAT2_CATALOGO (ID_CLIENTE,NOMBRE,FECHA_CREACION,RUTA_RAIZ) VALUES('".$idCliente."','".
-																						 $nombre."','".
-																						 date('Y-m-d H:i:s')."','cat/".
-																						 $idCliente.'_'.$nombre."/')";
+		$sqlE="INSERT INTO CAT2_CATALOGO (ID_CLIENTE,NOMBRE,FECHA_CREACION,RUTA_RAIZ) VALUES('".$idCliente."','".$nombre."','".date('Y-m-d H:i:s')."','cat/".$idCliente.'_'.$nombre."/')";
 		$resE=$objDb->sqlQuery($sqlE);
-		
 		if($resE){
 		  $this->crearCarpetasCatalogo($idCliente,$nombre);
 		  $res_x = 1;	
@@ -110,7 +111,7 @@ class catalogos{
 		}else{
 		  $res_x = 0;	
 		}
-      echo $res_x ;
+      	echo $res_x ;
 	}
   /**
 	*@method 		crearCarpetasCatalogo
@@ -119,11 +120,11 @@ class catalogos{
 	*
 	*/
  	public function crearCarpetasCatalogo($idCliente,$nombre){
-	 mkdir('cat/'.$idCliente.'_'.$nombre, 0700);
-		 mkdir('cat/'.$idCliente.'_'.$nombre.'/1', 0700);
-		 mkdir('cat/'.$idCliente.'_'.$nombre.'/2', 0700);
-		 mkdir('cat/'.$idCliente.'_'.$nombre.'/3', 0700);
-		 mkdir('cat/'.$idCliente.'_'.$nombre.'/4', 0700);	 			 
+	 	mkdir('cat/'.$idCliente.'_'.$nombre, 0777);
+		mkdir('cat/'.$idCliente.'_'.$nombre.'/1', 0777);
+		mkdir('cat/'.$idCliente.'_'.$nombre.'/2', 0777);
+		mkdir('cat/'.$idCliente.'_'.$nombre.'/3', 0777);
+		mkdir('cat/'.$idCliente.'_'.$nombre.'/4', 0777);	 			 
      }
    /**
 	*@method 		crearCarpetasCatalogo
